+++ source/StructureGraphLib/StructureGraph.cpp
@@ -2402,12 +2402,12 @@ void Graph::setAllControlPoints(Array2D_Vector3 all_points)
     int idx = 0;
     for(Structure::Node * n : nodes)
     {
-		if (idx > all_points.size() - 1) continue; // skip invalid assignments
+		if (idx > (int) all_points.size() - 1) continue; // skip invalid assignments
 
         auto expected = n->numCtrlPnts();
         auto got = all_points[idx].size();
 
-        if(expected != got) continue; // skip invalid assignments
+        if(expected != (int) got) continue; // skip invalid assignments
 
         n->setControlPoints(all_points[idx++]);
     }
